refactor: ui component files
